requirements.txt
python-telegram-bot==20.*
aiohttp
python-dotenv
pyTelegramBotAPI
Flask
gunicorn
chapa

